add new test methods
